juliareposetup

Functions

juliareposetup

Documentation for juliareposetup.

Functions

julia_repo_setup.binom — Method

Here's some inline maths: $\sqrt[n]{1 + x + x^2 + \ldots}$.

Here's an equation:

$\frac{n!}{k!(n - k)!} = \binom{n}{k}$

This is the binomial coefficient.

This is a paragraph.

function func(x)
    # ...
end

Another paragraph.

A code block without a "language":

function func(x)
    # ...
end

and another one with the "language" specified as julia:

function func(x)
    # ...
end

link to Example.jl Documentation

link to OpenCore Legacy Patcher

A paragraph containing a numbered footnote [1] and a named one [note].

Here's a quote:

Julia is a high-level, high-performance dynamic programming language for technical computing, with syntax that is familiar to users of other technical computing environments.

Another list:

item one

item two
f(x) = x

And a sublist:

sub-item one

sub-item two

\[f(a) = \frac{1}{2\pi}\int_{0}^{2\pi} (\alpha+R\cos(\theta))d\theta\]

Column One	Column Two	Column Three

Row 1	Column 2

Row 2	Row 2	Column 3

Note

This is the content of the note.

Tip

This is the content of the tip.

Beware!

And this is another one.

This warning admonition has a custom title: "Beware!".

julia_repo_setup.bar

julia_repo_setup.binom

julia_repo_setup.cosine

julia_repo_setup.f

julia_repo_setup.hello

julia_repo_setup.plus

julia_repo_setup.sine

julia_repo_setup.bar — Method

bar(x[, y])

Compute the Bar index between x and y.

If y is unspecified, compute the Bar index between all pairs of columns of x.

Examples

julia> bar([1, 2], [1, 2])
1

julia_repo_setup.cosine — Method

cosine(x)

Calculate the cosine of an angle x given in radians.

julia_repo_setup.f — Method

...

julia_repo_setup.hello — Method

hello(who::String)

Return "Hello, who".

julia> hello("Stranger")
"Hello, Stranger"

julia_repo_setup.plus — Method

plus(x, y)

Calculate add x and y.

julia_repo_setup.sine — Method

sine(x)

Calculate the sine of an angle x given in radians.

1Numbered footnote text.

note
Named footnote text containing several toplevel elements.

item one

item two

item three

function func(x)
    # ...
end